# backend/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: GTW IOT API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
paths:
  /asset/getAssetByUser:
    # binds a127 app logic to a route
    x-swagger-router-controller: asset
    get:
      tags:
      - asset
      summary: Retrieve list of assets
      description: Return list of assets by user id
      # used as the method name of the controller
      operationId: getAssetByUser
      parameters:
        - name: userID
          in: query
          description: input user ID
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            type: object
            properties:
              Items:
                type: array
                items:
                  $ref: "#/definitions/AssetModel"
              Count:
                type: integer
        # responses may fall through to errors
        400:
          description: Invalid Input
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Item not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /asset/getAssetAttributes:
    # binds a127 app logic to a route
    x-swagger-router-controller: asset
    get:
      tags:
      - asset
      description: Return list of asset's attributes for a given AssetID
      # used as the method name of the controller
      operationId: getAssetAttributes
      parameters:
        - name: AssetID
          in: query
          description: AssetID 
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/AssetModel"
        # responses may fall through to errors
        400:
          description: Invalid Input
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Item not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /asset/createAsset:
    x-swagger-router-controller: asset
    post:
      tags:
      - asset
      description: Create a new asset and add it to User list
      # used as the method name of the controller
      operationId: createAsset
      parameters:
      - in: body
        name: AssetObject
        description: Asset object thats needs to be added into table
        required: true
        schema:
          $ref: '#/definitions/AssetModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /asset/updateAsset:
    x-swagger-router-controller: asset
    put:
      tags:
      - asset
      summary: update an existing Asset
      description: update an existing Asset 
      # used as the method name of the controller
      operationId: updateAsset
      parameters:
      - in: body
        name: body
        description: Asset object thats needs to be updated
        required: true
        schema:
          $ref: '#/definitions/AssetModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /asset/deleteAsset:
    x-swagger-router-controller: asset
    delete:
      tags:
      - asset
      description: delete an existing Asset 
      # used as the method name of the controller
      operationId: deleteAsset
      parameters:
      - name: AssetID
        in: query
        required: true
        type: string
      - name: UserID
        in: query
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /user/updateUser:
    x-swagger-router-controller: userManage
    put:
      tags:
      - user
      summary: update a user
      description: update a user
      # used as the method name of the controller
      operationId: updateUser
      parameters:
      - in: body
        name: body
        description: User object thats need to be updated
        required: true
        schema:
          $ref: '#/definitions/UserModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  AssetModel:
    type: object
    properties:
      UserID:
        type: string
      AssetID:
        type: string
      AddTimeStamp:
        type: integer
      DisplayName:
        type: string
      LastestTimeStamp:
        type: integer
...
  UserModel:
    type: object
    properties:
      UserID:
        type: string
      Active:
        type: number
      EmailAddress:
        type: string
      Password:
        type: string
...
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string